Improve product list display.
Instead of loading the complete list of products at the beginning, we take chunks of products. Each time the user move the belt, a new group of products is added to the menu list.
After the original list of products is completely displayed, new products are randomly generated and added to the original list.